:recycle: Refactored the function that points to the position of the slide into a custom hook.
